+++ insanic/app.py
@@ -1,17 +1,19 @@
 import os
 import string
 
-from prometheus_client import Counter
 from sanic import Sanic
 from sanic.views import CompositionView
 from sanic_useragent import SanicUserAgent
 
+from insanic import __version__
 from insanic.conf import settings
 from insanic.functional import empty
 from insanic.handlers import ErrorHandler
+from insanic.metrics import InsanicMetrics
 from insanic.monitor import blueprint_monitor
 from insanic.log import get_logging_config, error_logger, logger
 from insanic.protocol import InsanicHttpProtocol
+from insanic.scopes import get_my_ip
 
 LISTENER_TYPES = ("before_server_start", "after_server_start", "before_server_stop", "after_server_stop")
 MIDDLEWARE_TYPES = ('request', 'response')
@@ -20,13 +22,9 @@ MIDDLEWARE_TYPES = ('request', 'response')
 class Insanic(Sanic):
     database = None
     _public_routes = empty
-    metrics = {}
+    metrics = empty
 
     def __init__(self, name, router=None, error_handler=None, app_config=()):
-
-        if "request_count" not in self.metrics:
-            self.metrics['request_count'] = Counter('request_count',
-                                                    'The number of requests this application has handled')
 
         if error_handler is None:
             error_handler = ErrorHandler()
@@ -84,9 +82,19 @@ class Insanic(Sanic):
                 else:
                     raise
 
-        settings.SERVICE_VERSION = _service_version()
+        service_version = _service_version()
+        settings.SERVICE_VERSION = service_version
         logger.info(f"{settings.SERVICE_NAME} v{settings.SERVICE_VERSION} service loaded.")
         self.attach_plugins()
+
+        self.metrics = InsanicMetrics
+        self.metrics.META.info({
+            "service": service_name,
+            "service_version": service_version,
+            "status": "OK",
+            "insanic_version": __version__,
+            "ip": get_my_ip()
+        })
 
     def attach_plugins(self):
         SanicUserAgent.init_app(self)