+++ insanic/middleware.py
@@ -4,6 +4,9 @@ from insanic.conf import settings
 
 
 def request_middleware(request):
-    request.app.metrics['request_count'].inc()
+    try:
+        request.app.metrics.REQUEST_COUNT.inc()
+    except AttributeError:
+        pass
     aiotask_context.set(settings.TASK_CONTEXT_CORRELATION_ID,
                         request.id)